webapp/config.py
import os

from dotenv import load_dotenv

load_dotenv()

basedir = os.path.abspath(os.path.dirname(__file__))

SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URI')

SQLALCHEMY_TRACK_MODIFICATIONS = False
SECRET_KEY = os.getenv('SECRET_KEY')
UPLOADED_PATH = os.path.join(basedir, 'uploads')
STATIC_PATH = os.path.join(basedir, 'static')
