requirements.txt
alembic==1.5.8
atomicwrites==1.4.0
attrs==20.3.0
Babel==2.9.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
fdfgen==0.16.1
flake8==3.9.1
flake8-bugbear==21.4.3
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
flake8-variables-names==0.0.4
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
pdfkit==0.6.1
pdfminer.six==20160202
pep8-naming==0.11.1
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
reportlab==3.5.67
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-SQLAlchemy==0.2
